feat: Recreate a common website section
My thought-process: since we are using flex, any
centering should be done in flex. So I made the
title a flex item as well, essentially nesting
flex containers. One flex container is vertical,
its flex items are the title and the item row. the
other flex container is a horizontal, for the plant
items (the item row). Each plant item is a div so
that its pic and text is together.