fix: Reconcile with provided solution
Removed the info-section class container that was
holding the title and the plants. Previously, this
container was used to make a flex column in order
to center the title and plants. However, this is
overly complicated as we can center the title by
just using text-align: center. Simpler is
definitely best, and I think I just need to
remember and become more familiar with CSS so that
 I can choose more cleaner solutions.

32px space between the title and the plants was
changed to be a margin-bottom on the title. Was
previously a gap: 32px due to we were using a flex
column container before, as mentioned in the
previous paragraph.

text-align: center was moved to the body selector
since it applies to both the plant-card text and
the title.

Added justify-content: center to the plants
selector. Previously had padding: 0 50px. LOl I
was just accounting for the space on the left and
right of the plants container. I guess I didn't
stretch the window enough to see that it wasn't
centered?? Or actually I think it was centered
because the plants container was previously a flex
item of another flex container (column, with the
title as its sibling flex item). And that parent
container had align-items: center; centering both
the title and this plants container. So that's prob
what it was