fix: Reconcile with provided solution
In my solution I used an additional flex container
than the provided solution used. After reviewing,
I realized that it was similar to what I had done
in the previous exercise 04-flex-information,
where there were some areas that did not need a
flex container or flex properties and could just
be solved with margin or padding instead. In this
exercise, I used the additional flex container to
add spacing (using gap) between the header and the
 text and between the text and the buttons at the
 bottom. The solution solved this by just using
 margin-bottom.

 I drew out the diagrams for my solution and the
 provided solution, and I realized I had not
 considered that we could add a flex container
 inside a flex item even when that flex item was
 not a flex container itself. For some reason, I
 just thought flex containers needed to be inside
 of other flex containers. But I tested this out
 and it is possible: a flex item (not a flex
 container at all) can contain a flex container,
 and that flex container can have its own flex
 items.